Move blob_index.h to db/
Summary:
Extracted from PR #5903 for technical reasons.

Test Plan:
make check
